Move startup code into a more appropriate position
