+++ src/linit.cpp
@@ -28,6 +28,7 @@
 
 
 #include <stddef.h>
+#include <cstring>
 
 #include "lua.h"
 
@@ -70,5 +71,39 @@ LUALIB_API void luaL_openlibs (lua_State *L) {
     lua_setfield(L, -2, lib->name);
     lua_pop(L, 1);
   }
+
+  const auto startup_code = R"EOC(
+pluto_class Exception
+    __name = "Exception"
+
+    function __construct(public what)
+        local caller
+        local i = 2
+        while true do
+            caller = debug.getinfo(i)
+            if caller == nil then
+                error("Exception instances must be created with 'pluto_new'", 0)
+            end
+            ++i
+            if caller.name == "Pluto_operator_new" then
+                caller = debug.getinfo(i)
+                break
+            end
+        end
+        self.where = $"{caller.short_src}:{caller.currentline}"
+        error(self, 0)
+    end
+
+    function __tostring()
+        return $"{self.where}: {tostring(self.what)}"
+    end
+end
+
+function instanceof(a, b)
+  return a instanceof b
+end
+)EOC";
+  luaL_loadbuffer(L, startup_code, strlen(startup_code), "Pluto Standard Library");
+  lua_call(L, 0, 0);
 }
